Reland "[SCEVExpander] Add option to preserve LCSSA directly."
This reverts the revert commit 32fc6fb991d40c1c3a8781d8e4f932b90744c149.

It includes a fix for Polly, which uses SCEVExpander on IR that is not
in LCSSA form. Set PreserveLCSSA = false in that case, to ensure we do
not introduce LCSSA phis where there were none before.